// src/interfaces/ISzolgaltatasok.ts
export interface ISzolgaltatasok {
    id: number;
    title: string;
    href: string;
    description: string;
    iconClass: string;
    colorRed: number;
    colorGreen: number;
    colorBlue: number;
    bulletpoints: string[];
    more: string[];
    isReszletek: boolean;
    reszletek: string;
}